fix: revamp namespaces to make room for Spout
