+++ phpstan.neon.dist
@@ -23,4 +23,4 @@ parameters:
         - '#Call to deprecated method registerUniqueLoader\(\) of class Doctrine\\Common\\Annotations\\AnnotationRegistry\.#'
 
         # false positives
-        - '#PHPDoc tag @throws with type Xezilaires\\Infrastructure\\Symfony\\Serializer\\Exception is not compatible with parent#'
+        - '#PHPDoc tag @throws with type Xezilaires\\Bridge\\Symfony\\Serializer\\Exception is not compatible with parent#'
