+++ src/Xezilaires/Bridge/Symfony/Command/SerializeCommand.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Xezilaires\Infrastructure\Symfony\Command;
+namespace Xezilaires\Bridge\Symfony\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -22,9 +22,9 @@ use Symfony\Component\Serializer\Encoder\CsvEncoder;
 use Symfony\Component\Serializer\Encoder\JsonEncode;
 use Symfony\Component\Serializer\Encoder\XmlEncoder;
 use Symfony\Component\Serializer\Serializer;
-use Xezilaires\Infrastructure\Symfony\Serializer\ObjectNormalizer;
+use Xezilaires\Bridge\PhpSpreadsheet\Iterator;
+use Xezilaires\Bridge\Symfony\Serializer\ObjectNormalizer;
 use Xezilaires\Metadata\Annotation\AnnotationDriver;
-use Xezilaires\PhpSpreadsheetIterator;
 
 /**
  * Class SerializeCommand.
@@ -88,7 +88,7 @@ class SerializeCommand extends Command
         }
 
         $driver = new AnnotationDriver();
-        $iterator = new PhpSpreadsheetIterator(new \SplFileObject($path), $driver->getMetadataMapping($class, ['reverse' => $reverse]));
+        $iterator = new Iterator(new \SplFileObject($path), $driver->getMetadataMapping($class, ['reverse' => $reverse]));
         $serializer = new Serializer($normalizers, $encoders);
         $output->write($serializer->serialize($iterator, $format));
