+++ tests/Xezilaires/FilterIteratorTest.php
@@ -11,14 +11,13 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Xezilaires\Test\Infrastructure\Utility;
+namespace Xezilaires\Test;
 
 use PHPUnit\Framework\TestCase;
-use Xezilaires\Infrastructure\Utility\FilterIterator;
-use Xezilaires\Test\FakeIterator;
+use Xezilaires\FilterIterator;
 
 /**
- * @covers \Xezilaires\Infrastructure\Utility\FilterIterator
+ * @covers \Xezilaires\FilterIterator
  */
 class FilterIteratorTest extends TestCase
 {
